Update documentation to match defined glossary
Updated remaining documentation to match the defined glossary under
`glossary/main.rst` file.

Updated `aws rds start-port-forwarding-session` command's documentation
to include a warning about the prerequisites before being able to use
that command.

Updated defined glossary's titles to match the exact terms we are using
in the rest of the documentation. Improved also some message to have a
longer version of the changed words (BE->Backend and so on).

Added AWS prefix term to all the mentioned AWS services to avoid
confusions.