Apply HOP Application term and AWS prefix in a doc
We needed to replace `project` word by `HOP Application` and
`application` terms, depending on the case, as that is now how we
reference to those terms, according to the Glossary we have defined.

Also added AWS prefix to each service/feature mentioned in the document,
as we are doing in other places (not yet in all of them).