Update docs to include a bootstrap command alias
We have made `bootstrap copy-settings` an alias of `bootstrap
create-settings-file` command, as we think the new name is more
descriptive for the users. We still want to keep the alias as the older
version of the command until we make sure there are no longer
references to it (video-tutorials and so on.).