Add a requirements file for Sphinx packages
We are using some additional Sphinx packages (like the RTD theme), in
the local builds. If we want to use those same packages in the builds
triggered by the ReadTheDocs CI/CD builds, we need to make sure they
are installed as part of the build setup.

Once we need to add that file for the CI/CD builds, we may as well use
it to install the same packages in the local (dockerized) build system.